data-pipeline-assignment/requirements.txt
apache-airflow==2.7.3
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary==2.9.9
lxml==4.9.3
pandas==2.1.3
